64-bit addressing support -Switched all 32-bit unsigned integers in VkFFT configuration to 64-bit.
!!!You have to change your code if you passed buffers by pointer there!!!

-VkFFT now supports addressing of buffers spanning more than 4GB in memory. Enabled automatically when needed, can be forced by useUint64 option. Tested validity on 4000x1000x1000 3D FFT system on AMD MI100 (almost 32GB of used GPU memory)
-64-bit addressing works in CUDA, HIP and OpenCL backends. No physical addressing mode in Vulkan shaders, so use multiple-buffer support to achieve emulation of physical addressing there
-This is the first step towards multi-GPU support in VkFFT
-Fixed some compiler warnings

+++ cufft_scripts/precision_cuFFT.h
@@ -1,1 +1,1 @@
-void launch_precision_cuFFT_single(void* inputC, void* output_cuFFT, uint32_t* dims);
+void launch_precision_cuFFT_single(void* inputC, void* output_cuFFT, uint64_t* dims);
