+++ cufft_scripts/precision_cuFFT_double.h
@@ -1,1 +1,1 @@
-void launch_precision_cuFFT_double(void* inputC, void* output_cuFFT, uint32_t* dims);
+void launch_precision_cuFFT_double(void* inputC, void* output_cuFFT, uint64_t* dims);
