+++ cufft_scripts/precision_cuFFT_r2c.h
@@ -1,1 +1,1 @@
-void launch_precision_cuFFT_r2c(void* inputC, void* output_cuFFT, uint32_t* dims);
+void launch_precision_cuFFT_r2c(void* inputC, void* output_cuFFT, uint64_t* dims);
