+++ rocfft_scripts/precision_rocFFT_double.h
@@ -1,1 +1,1 @@
-void launch_precision_rocFFT_double(void* inputC, void* output_cuFFT, uint32_t* dims);
+void launch_precision_rocFFT_double(void* inputC, void* output_cuFFT, uint64_t* dims);
